# openapi.yaml
openapi: 3.1.0
info:
  title: Incident Tracker API
  version: 0.3.0
  description: >-
    Multi-tenant incident tracker API with session auth, tenancy, and incident workflows.
    ID-based routes return 404 for objects that are missing or inaccessible in the active
    tenant context, and return 403 only when object existence is established but permission
    is denied.
servers:
  - url: http://localhost:3000
paths:
...
  /v1/auth/login:
    post:
      summary: Login with email and password
      operationId: login
      security:
        - CsrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStateAuthenticatedResponse'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: CSRF validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: Account temporarily locked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/tenants:
    get:
      summary: List tenants for the authenticated user
      operationId: listTenants
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Tenant memberships for the current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantListResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a tenant and owner membership
      operationId: createTenant
      security:
        - SessionCookie: []
          CsrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTenantResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/audit-logs:
    get:
      summary: List tenant audit logs
      operationId: listAuditLogs
      security:
        - SessionCookie: []
        - ApiKeyHeader: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Audit log page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '400':
          description: Invalid tenant context or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tenant not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/usage:
    get:
      summary: Read tenant usage and quota summary
      operationId: getUsageSummary
      security:
        - SessionCookie: []
        - ApiKeyHeader: []
      responses:
        '200':
          description: Usage summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageSummaryResponse'
        '400':
          description: Invalid tenant context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tenant not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/incidents:
    get:
      summary: List incidents for active tenant
      operationId: listIncidents
      security:
        - SessionCookie: []
        - ApiKeyHeader: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Incident page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentListResponse'
        '400':
          description: Invalid tenant context or pagination cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Active tenant membership not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create incident
      operationId: createIncident
      security:
        - SessionCookie: []
          CsrfToken: []
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentRequest'
      responses:
        '201':
          description: Incident created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentItemResponse'
        '400':
          description: Invalid tenant context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Tenant write quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Active tenant membership not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: incident_tracker_sid
    CsrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
...
    CreateTenantRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 80
...
    CreateIncidentRequest:
      type: object
      required: [title, severity, startTime, impactedServices]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        description:
          type: string
          maxLength: 4000
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        startTime:
          type: string
          format: date-time
        impactedServices:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 100
...
    OrgRole:
      type: string
      enum: [Owner, Admin, Responder, Viewer, Billing]
    IncidentSeverity:
      type: string
      enum: [SEV1, SEV2, SEV3, SEV4]
    IncidentStatus:
      type: string
      enum: [declared, investigating, mitigating, monitoring, resolved, closed]
...
    AuthType:
      type: string
      enum: [session, api_key]
...
    Tenant:
      type: object
      required: [id, name, slug, createdByUserId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        createdByUserId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Membership:
      type: object
      required: [id, tenantId, userId, role, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/OrgRole'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Incident:
      type: object
      required:
        [
          id,
          tenantId,
          title,
          description,
          severity,
          status,
          startTime,
          endTime,
          declaredByUserId,
          impactedServices,
          createdAt,
          updatedAt
        ]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        startTime:
          type: string
          format: date-time
        endTime:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        declaredByUserId:
          type: string
          format: uuid
        impactedServices:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AuthenticatedUser:
      type: object
      required: [id, email, passwordUpdatedAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        passwordUpdatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    AuthStateAuthenticatedResponse:
      type: object
      required: [authenticated, authType, user, csrfToken]
      properties:
        authenticated:
          type: boolean
          const: true
        authType:
          $ref: '#/components/schemas/AuthType'
        user:
          $ref: '#/components/schemas/AuthenticatedUser'
        csrfToken:
          type: string
    TenantListItem:
      type: object
      required: [tenant, membership]
      properties:
        tenant:
          $ref: '#/components/schemas/Tenant'
        membership:
          type: object
          required: [id, role]
          properties:
            id:
              type: string
              format: uuid
            role:
              $ref: '#/components/schemas/OrgRole'
    TenantListResponse:
      type: object
      required: [activeTenantId, tenants, csrfToken]
      properties:
        activeTenantId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        tenants:
          type: array
          items:
            $ref: '#/components/schemas/TenantListItem'
        csrfToken:
          type: string
    CreateTenantResponse:
      type: object
      required: [tenant, membership, activeTenantId, csrfToken]
      properties:
        tenant:
          $ref: '#/components/schemas/Tenant'
        membership:
          $ref: '#/components/schemas/Membership'
        activeTenantId:
          type: string
          format: uuid
        csrfToken:
          type: string
...
    IncidentListResponse:
      type: object
      required: [incidents, nextCursor, csrfToken]
      properties:
        incidents:
          type: array
          items:
            $ref: '#/components/schemas/Incident'
        nextCursor:
          oneOf:
            - type: string
            - type: 'null'
        csrfToken:
          type: string
    IncidentItemResponse:
      type: object
      required: [incident, csrfToken]
      properties:
        incident:
          $ref: '#/components/schemas/Incident'
        csrfToken:
          type: string
...
    AuditLogEvent:
      type: object
      required:
        [id, tenantId, actorUserId, action, targetType, targetId, metadata, traceId, ipAddress, userAgent, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        actorUserId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        action:
          type: string
        targetType:
          oneOf:
            - type: string
            - type: 'null'
        targetId:
          oneOf:
            - type: string
            - type: 'null'
        metadata:
          type: object
          additionalProperties: true
        traceId:
          oneOf:
            - type: string
            - type: 'null'
        ipAddress:
          oneOf:
            - type: string
            - type: 'null'
        userAgent:
          oneOf:
            - type: string
            - type: 'null'
        createdAt:
          type: string
          format: date-time
    AuditLogListResponse:
      type: object
      required: [events, nextCursor, csrfToken]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEvent'
        nextCursor:
          oneOf:
            - type: string
            - type: 'null'
        csrfToken:
          type: string
    UsageSummary:
      type: object
      required: [tenantId, metric, windowHours, used, limit, remaining]
      properties:
        tenantId:
          type: string
          format: uuid
        metric:
          type: string
        windowHours:
          type: integer
          minimum: 1
        used:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 0
        remaining:
          type: integer
          minimum: 0
    UsageSummaryResponse:
      type: object
      required: [usage, csrfToken]
      properties:
        usage:
          $ref: '#/components/schemas/UsageSummary'
        csrfToken:
          type: string
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        details:
          description: Optional structured metadata for debugging and validation.